restructure the multi modality filers
