+++ rust/sglang-server/src/utils/runtime.rs
@@ -51,7 +51,7 @@ pub struct Runtime {
     pub tokenizer: Option<Arc<dyn tokenizer::TextTokenizer>>,
     /// MM results parked between a worker's `MmEncoded` and the scheduler drain
     /// (`Server.take_mm`).
-    pub mm_sidecar: crate::mm::Sidecar,
+    pub mm_sidecar: crate::multi_modality::Sidecar,
     /// Worker join handles, joined by `request_shutdown` / `Drop`.
     threads: Mutex<Vec<JoinHandle<()>>>,
     /// The single shutdown sender.
@@ -71,10 +71,10 @@ impl Runtime {
     /// MM preprocessing floats over that whole set (rather than owning cores
     /// that idle between bursts) and never preempts the scheduler's reserved
     /// cores.
-    pub fn spawn_mm_pool(&self, workers: usize, ctx: Arc<crate::mm::Context>) {
+    pub fn spawn_mm_pool(&self, workers: usize, ctx: Arc<crate::multi_modality::Context>) {
         let mut threads = self.threads.lock().unwrap();
         spawn_pool("mm-worker", None, workers.max(1), &mut threads, |_| {
-            crate::mm::MmWorker::new(self.mm.clone(), self.tm.clone(), ctx.clone())
+            crate::multi_modality::MmWorker::new(self.mm.clone(), self.tm.clone(), ctx.clone())
         });
     }
 
@@ -180,7 +180,7 @@ pub fn start(cfg: RuntimeConfig) -> Result<Runtime, String> {
         .map(|t| Arc::new(tokenizer::DynamoTokenizer::new(t.clone())) as _);
 
     // Shared: MM workers park, the Python drain pops, tm-ingress purges.
-    let mm_sidecar: crate::mm::Sidecar = Default::default();
+    let mm_sidecar: crate::multi_modality::Sidecar = Default::default();
 
     // --- Detokenizer shards (pinned, CPU bound) ---
     {
